+++ packages/tsurlfilter/test/engine/cosmetic-engine/cosmetic-engine.test.ts
@@ -317,19 +317,97 @@ describe('Test cosmetic engine - JS rules', () => {
         expect(scriptletData.func.name).toBe('log');
     });
 
-    it('checks scriptlet exceptions', () => {
-        const ruleContent = "//scriptlet('abort-on-property-read', 'I10C')";
-        const jsRule = `testcases.adguard.com,surge.sh#%#${ruleContent}`;
-        const jsExceptionRule = `testcases.adguard.com,surge.sh#@%#${ruleContent}`;
-        const cosmeticEngine = new CosmeticEngine(createTestRuleStorage(1, [
-            jsRule,
-            jsExceptionRule,
-        ]));
+    describe('scriptlets exceptions', () => {
+        it('checks scriptlet exceptions', () => {
+            const ruleContent = "//scriptlet('abort-on-property-read', 'I10C')";
+            const jsRule = `testcases.adguard.com,surge.sh#%#${ruleContent}`;
+            const jsExceptionRule = `testcases.adguard.com,surge.sh#@%#${ruleContent}`;
+            const cosmeticEngine = new CosmeticEngine(createTestRuleStorage(1, [
+                jsRule,
+                jsExceptionRule,
+            ]));
 
-        const result = cosmeticEngine.match(createRequest('testcases.adguard.com'), CosmeticOption.CosmeticOptionAll);
+            const result = cosmeticEngine.match(
+                createRequest('testcases.adguard.com'),
+                CosmeticOption.CosmeticOptionAll,
+            );
 
-        expect(result.JS.specific.length).toBe(0);
-        expect(result.JS.generic.length).toBe(0);
+            expect(result.JS.specific.length).toBe(0);
+            expect(result.JS.generic.length).toBe(0);
+        });
+
+        it('checks scriptlet exception even if content has different quotes', () => {
+            const jsRule = String.raw`testcases.adguard.com,surge.sh#%#//scriptlet('abort-on-property-read', 'I10\'C')`;
+            // eslint-disable-next-line max-len
+            const jsExceptionRule = String.raw`testcases.adguard.com,surge.sh#@%#//scriptlet("abort-on-property-read", "I10'C")`;
+            const cosmeticEngine = new CosmeticEngine(createTestRuleStorage(1, [
+                jsRule,
+                jsExceptionRule,
+            ]));
+
+            const result = cosmeticEngine.match(
+                createRequest('testcases.adguard.com'),
+                CosmeticOption.CosmeticOptionAll,
+            );
+
+            expect(result.JS.specific.length).toBe(0);
+            expect(result.JS.generic.length).toBe(0);
+        });
+
+        it('generic #@%#//scriptlet() disables all scriptlet rules', () => {
+            const allowlistScriptletRule = '#@%#//scriptlet()';
+            const specificScriptletRule = "example.org#%#//scriptlet('set-cookie', 'adcook2', '2')";
+            const genericScriptletRule = "#%#//scriptlet('set-local-storage-item', 'aditem1', '1')";
+            const cosmeticEngine = new CosmeticEngine(createTestRuleStorage(1, [
+                allowlistScriptletRule,
+                specificScriptletRule,
+                genericScriptletRule,
+            ]));
+
+            const result = cosmeticEngine.match(createRequest('example.org'), CosmeticOption.CosmeticOptionAll);
+
+            expect(result.JS.specific.length).toBe(0);
+            expect(result.JS.generic.length).toBe(0);
+        });
+
+        it('specific #@%#//scriptlet() disables all scriptlet rules only matching request', () => {
+            const allowlistScriptletRule = 'example.org#@%#//scriptlet()';
+            const specificScriptletRule = "example.org#%#//scriptlet('set-cookie', 'adcook2', '2')";
+            const specificScriptletRule2 = "example.com#%#//scriptlet('set-cookie', 'adcook2', '2')";
+            const genericScriptletRule = "#%#//scriptlet('set-local-storage-item', 'aditem1', '1')";
+            const cosmeticEngine = new CosmeticEngine(createTestRuleStorage(1, [
+                allowlistScriptletRule,
+                specificScriptletRule,
+                genericScriptletRule,
+                specificScriptletRule2,
+            ]));
+
+            const result = cosmeticEngine.match(createRequest('example.org'), CosmeticOption.CosmeticOptionAll);
+            expect(result.JS.specific.length).toBe(0);
+            expect(result.JS.generic.length).toBe(0);
+
+            const result2 = cosmeticEngine.match(createRequest('example.com'), CosmeticOption.CosmeticOptionAll);
+            expect(result2.JS.specific.map((r) => r.getText())).toEqual([specificScriptletRule2]);
+            expect(result2.JS.generic.map((r) => r.getText())).toEqual([genericScriptletRule]);
+        });
+
+        it('allowlists scriptlets with the same name', () => {
+            const allowlistScriptletRule = "#@%#//scriptlet('set-cookie')";
+            const specificSetCookieScriptletRule = "example.org#%#//scriptlet('set-cookie', 'test1', '1')";
+            const genericSetCookieScriptletRule = "#%#//scriptlet('set-cookie', 'test2', '2')";
+            const specificSetStorageScriptletRule = "example.org#%#//scriptlet('set-local-storage-item', 'test1', '1')";
+            const genericSetStorageScriptletRule = "#%#//scriptlet('set-local-storage-item', 'test2', '2')";
+            const cosmeticEngine = new CosmeticEngine(createTestRuleStorage(1, [
+                allowlistScriptletRule,
+                specificSetCookieScriptletRule,
+                genericSetCookieScriptletRule,
+                specificSetStorageScriptletRule,
+                genericSetStorageScriptletRule,
+            ]));
+            const result = cosmeticEngine.match(createRequest('example.org'), CosmeticOption.CosmeticOptionAll);
+            expect(result.JS.specific.map((r) => r.getText())).toEqual([specificSetStorageScriptletRule]);
+            expect(result.JS.generic.map((r) => r.getText())).toEqual([genericSetStorageScriptletRule]);
+        });
     });
 });